Working module manager
Every module has an enum moduleType. Either Switch, Dpad, ThreeButtons or OneButton. When the current module of a type finishes the next one is loaded (if there is a next one).

The module class has a LoadVisuals() method which should be overriden for each module. Most of the time it will just be adding the visual as a child.

The module class has StartModule() and StartTimer() methods. StartModule calls the StartTimer and LoadVisuals methods. Without calling StartModule() a module appear or function on screen. The ModuleManager class handles the calling of StartModule() for each module.